rename api to fetchdata.ts
