+++ frontend/react/src/pages/Login/LoginView.tsx
@@ -3,8 +3,7 @@ import { useNavigate, useParams } from "react-router-dom";
 import { useContext, useState } from "react";
 import { UserContext } from "../../App";
 import { FaPortrait, FaLock } from "react-icons/fa";
-import { getClient } from "../../api/API";
-
+import { getClient } from "../../api/FetchData";
 
 const Login = () => {
   const { userId } = useContext(UserContext);
@@ -22,17 +21,17 @@ const Login = () => {
 
   const fetchClient = async () => {
     try {
-        const response = await getClient(userId)
+      const response = await getClient(userId);
 
-       console.log(response?.reservation)
-        response?.reservation ? navigate(`/reservation/${response.id}`)
-            : navigate(`/vacancies/${response?.id}`)
-            console.log(response)
+      console.log(response?.reservation);
+      response?.reservation
+        ? navigate(`/reservation/${response.id}`)
+        : navigate(`/vacancies/${response?.id}`);
+      console.log(response);
     } catch (error) {
-        console.error(error);
+      console.error(error);
     }
-}
-
+  };
 
   const loginClient = async () => {
     await fetchClient();
